Extract TryClone trait into its own crate
As per the discussion at
https://github.com/informalsystems/tendermint-rs/pull/938#discussion_r677457303,
this extracts the `TryClone` trait into a new crate called
`tendermint-std-ext` in the `std-ext` directory.

This new crate is intended to contain any code that we need that extends
the Rust standard library.
