+++ p2p/Cargo.toml
@@ -43,6 +43,7 @@ zeroize = "1"
 # path dependencies
 tendermint = { path = "../tendermint", version = "0.21.0" }
 tendermint-proto = { path = "../proto", version = "0.21.0" }
+tendermint-std-ext = { path = "../std-ext", version = "0.21.0" }
 
 # optional dependencies
 prost-amino = { version = "0.6", optional = true }
