+++ std-ext/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name       = "tendermint-std-ext"
+version    = "0.21.0"
+edition    = "2018"
+license    = "Apache-2.0"
+homepage   = "https://www.tendermint.com/"
+repository = "https://github.com/informalsystems/tendermint-rs"
+readme     = "README.md"
+keywords   = ["blockchain", "cosmos", "tendermint"]
+categories = ["development-tools"]
+authors    = ["Informal Systems <hello@informal.systems>"]
+
+description = """
+    tendermint-std-ext contains extensions to the Rust standard library for use
+    from tendermint-rs.
+    """
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
