+++ std-ext/README.md
@@ -1,0 +1,24 @@
+[![Crate][crate-image]][crate-link]
+[![Docs][docs-image]][docs-link]
+
+See the [repo root] for build status, license, rust version, etc.
+
+# tendermint-std-ext
+
+Extensions to the [Rust standard library][std] for use by Tendermint in Rust.
+
+## Documentation
+
+See documentation on [crates.io][docs-link].
+
+[//]: # (badges)
+
+[crate-image]: https://img.shields.io/crates/v/tendermint-std-ext.svg
+[crate-link]: https://crates.io/crates/tendermint-std-ext
+[docs-image]: https://docs.rs/tendermint-std-ext/badge.svg
+[docs-link]: https://docs.rs/tendermint-std-ext/
+
+[//]: # (general links)
+
+[repo root]: https://github.com/informalsystems/tendermint-rs
+[std]: https://doc.rust-lang.org/std/
